# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/google/callback:
    get:
      operationId: googleAuthCallback
      tags: [calendar]
      summary: Google OAuth callback
      description: Handles OAuth callback from Google and stores tokens
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect to frontend with success/error
...